requirements.txt
beautifulsoup4
pylint
pytest
requests
psycopg2
python-dotenv

